requirements.txt
flask==3.1.2
pillow==8.4.0
sentencepiece==0.2.1
transformers==4.55.4 
fugashi==1.5.1
unidic-lite==1.0.8
accelerate==1.10.1
paddlepaddle==3.1.1
paddleocr==3.2.0
opencv-python-headless==4.12.0.88
